chore(project): 🔧 streamline workflows and editor setup
