+++ .github/workflows/release.yml
@@ -1,0 +1,58 @@
+name: Release
+
+on:
+  workflow_dispatch:
+    inputs:
+      version:
+        description: 'Version to release'
+        required: true
+        type: string
+
+env:
+  artifact: EsnyaResoniteModTemplate.dll
+  project_path: EsnyaResoniteModTemplate
+
+jobs:
+  release:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      - uses: actions/setup-dotnet@v4
+        with:
+          dotnet-version: '9.0.x'
+          cache: true
+          cache-dependency-path: |
+            Directory.Packages.props
+            **/*.csproj
+      - run: dotnet tool restore
+      - name: Set version
+        run: sed -i "s#<Version>[^<]*</Version>#<Version>${{ inputs.version }}</Version>#" Directory.Build.props
+      - run: dotnet csharpier format Directory.Build.props
+      - run: dotnet csharpier check .
+      - run: dotnet restore
+      - run: dotnet build --configuration Release --no-restore
+      - run: dotnet test --configuration Release --no-build
+      - name: Commit and tag
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
+          git commit -am "chore(release): 🔖 v${{ inputs.version }}"
+          git tag "v${{ inputs.version }}"
+          git push origin HEAD
+          git push origin "v${{ inputs.version }}"
+      - name: Detect prerelease
+        id: prerelease
+        run: |
+          if [[ "${{ inputs.version }}" == *-* ]]; then
+            echo "flag=true" >> $GITHUB_OUTPUT
+          else
+            echo "flag=false" >> $GITHUB_OUTPUT
+          fi
+      - uses: softprops/action-gh-release@v2
+        with:
+          files: ${{ env.project_path }}/bin/Release/net9.0/${{ env.artifact }}
+          tag_name: v${{ inputs.version }}
+          generate_release_notes: true
+          prerelease: ${{ steps.prerelease.outputs.flag }}
