+++ Directory.Build.props
@@ -1,6 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
 <Project>
-  <!-- Common Package Properties - Update these when creating a new mod from this template -->
   <PropertyGroup>
     <Authors>esnya</Authors>
     <Description>A Resonite mod created with EsnyaResoniteModTemplate</Description>
@@ -13,21 +11,17 @@
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <Version>0.0.0-template.1</Version>
   </PropertyGroup>
-  <!-- Common Build Properties -->
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
     <LangVersion>12.0</LangVersion>
     <Nullable>enable</Nullable>
     <Deterministic>true</Deterministic>
     <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-    <!-- Debug/Release specific settings -->
     <DebugType Condition="'$(Configuration)'=='Debug'">embedded</DebugType>
     <DebugType Condition="'$(Configuration)'=='Release'">pdbonly</DebugType>
     <Optimize Condition="'$(Configuration)'=='Release'">true</Optimize>
-    <!-- Assembly Metadata -->
     <AssemblyTitle>$(MSBuildProjectName)</AssemblyTitle>
     <Product>$(MSBuildProjectName)</Product>
-    <!-- Code Quality -->
     <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
     <EnableNETAnalyzers>true</EnableNETAnalyzers>
     <AnalysisLevel>latest-all</AnalysisLevel>
@@ -36,34 +30,27 @@
     <WarningsAsErrors />
     <WarningsNotAsErrors>NU1701</WarningsNotAsErrors>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <!-- Allow .NET Framework dependencies -->
     <CopyLocalLockFileAssemblies>$(IsTestProject)</CopyLocalLockFileAssemblies>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
   </PropertyGroup>
-  <!-- Resonite Path Discovery -->
   <PropertyGroup Condition="'$(ResonitePath)'==''">
-    <!-- Default to local folder -->
     <ResonitePath>$([System.IO.Path]::Combine('$(MSBuildThisFileDirectory)', 'Resonite'))</ResonitePath>
-    <!-- Windows Steam path -->
     <ResonitePath
       Condition="'$(OS)'=='Windows_NT' AND Exists('C:\\Program Files (x86)\\Steam\\steamapps\\common\\Resonite\\')"
       >C:\Program Files (x86)\Steam\steamapps\common\Resonite</ResonitePath
     >
-    <!-- Linux Steam path -->
     <ResonitePath
-      Condition="'$(OS)'!='Windows_NT' AND Exists('$(USERPROFILE)\.steam\steam\steamapps\common\Resonite\')"
-      >$(USERPROFILE)\.steam\steam\steamapps\common\Resonite</ResonitePath
+      Condition="'$(OS)'!='Windows_NT' AND Exists('$(USERPROFILE)\.steam\\steam\\steamapps\\common\\Resonite\\')"
+      >$(USERPROFILE)\.steam\\steam\\steamapps\\common\\Resonite</ResonitePath
     >
     <ResonitePath
       Condition="'$(OS)'!='Windows_NT' AND Exists('$(HOME)/.steam/steam/steamapps/common/Resonite/')"
       >$(HOME)/.steam/steam/steamapps/common/Resonite</ResonitePath
     >
   </PropertyGroup>
-  <!-- Normalize ResonitePath -->
   <PropertyGroup>
     <ResonitePath>$([MSBuild]::EnsureTrailingSlash('$(ResonitePath)'))</ResonitePath>
   </PropertyGroup>
-  <!-- Derived Resonite Paths -->
   <PropertyGroup>
     <ResoniteLibrariesPath>$([System.IO.Path]::Combine('$(ResonitePath)', 'Libraries'))</ResoniteLibrariesPath>
     <ResoniteLibsPath>$([System.IO.Path]::Combine('$(ResonitePath)', 'rml_libs'))</ResoniteLibsPath>
@@ -71,9 +58,7 @@
     <ResoniteHotReloadPath>$([System.IO.Path]::Combine('$(ResoniteModsPath)', 'HotReloadMods'))</ResoniteHotReloadPath>
     <ResoniteFrooxEnginePath>$([System.IO.Path]::Combine('$(ResonitePath)', 'FrooxEngine.dll'))</ResoniteFrooxEnginePath>
   </PropertyGroup>
-  <!-- Common Resonite References -->
   <ItemGroup>
-    <!-- Core Resonite Mod Dependencies -->
     <Reference Include="ResoniteModLoader">
       <HintPath>$([System.IO.Path]::Combine('$(ResoniteLibrariesPath)', 'ResoniteModLoader.dll'))</HintPath>
       <Private>$(IsTestProject)</Private>
@@ -82,13 +67,11 @@
       <HintPath>$([System.IO.Path]::Combine('$(ResoniteLibsPath)', '0Harmony.dll'))</HintPath>
       <Private>$(IsTestProject)</Private>
     </Reference>
-    <!-- Debug-only references -->
     <Reference Include="ResoniteHotReloadLib" Condition="'$(Configuration)'=='Debug'">
       <HintPath>$([System.IO.Path]::Combine('$(ResoniteLibsPath)', 'ResoniteHotReloadLib.dll'))</HintPath>
       <Private>$(IsTestProject)</Private>
     </Reference>
   </ItemGroup>
-  <!-- Resonite Engine References -->
   <ItemGroup Condition="Exists('$(ResoniteFrooxEnginePath)')">
     <Reference Include="FrooxEngine">
       <HintPath>$(ResoniteFrooxEnginePath)</HintPath>
@@ -98,7 +81,6 @@
   <ItemGroup Condition="!Exists('$(ResoniteFrooxEnginePath)')">
     <PackageReference Include="Resonite.GameLibs" />
   </ItemGroup>
-  <!-- Common Items -->
   <ItemGroup>
     <None Include="$(MSBuildThisFileDirectory).github\**\*" LinkBase=".github" />
     <EditorConfigFiles Remove="$(MSBuildProjectDirectory)**\*.editorconfig" />
