+++ Directory.Build.targets
@@ -1,30 +1,19 @@
-<?xml version="1.0" encoding="utf-8"?>
 <Project>
-  <!--
-    Install Target - Install mod to Resonite
-    Usage: dotnet build -t:Install
-
-    Prerequisites: Project must be built first (dotnet build)
-    Installs to main mods folder, and additionally to HotReload folder for Debug builds
-  -->
   <Target
     Name="Install"
     Inputs="$(TargetPath)"
-    Outputs="$(ResoniteModsPath)\$(TargetFileName);$(ResoniteHotReloadPath)\$(TargetFileName)"
+    Outputs="$(ResoniteModsPath)\\$(TargetFileName);$(ResoniteHotReloadPath)\\$(TargetFileName)"
   >
-    <!-- Check if build output exists -->
     <Error
       Text="Build output not found at $(TargetPath). Please run 'dotnet build' first."
       Condition="!Exists('$(TargetPath)')"
     />
-    <!-- Always install to main mods folder -->
     <Message
       Text="Installing $(TargetFileName) to Resonite mods folder: $(ResoniteModsPath)"
       Importance="high"
     />
     <MakeDir Directories="$(ResoniteModsPath)" Condition="!Exists('$(ResoniteModsPath)')" />
     <Copy SourceFiles="$(TargetPath)" DestinationFolder="$(ResoniteModsPath)" />
-    <!-- Additionally install to HotReload folder for Debug builds -->
     <Message
       Text="Installing $(TargetFileName) to Resonite hot reload folder: $(ResoniteHotReloadPath)"
       Importance="high"
