+++ Directory.Packages.props
@@ -1,25 +1,14 @@
-<?xml version="1.0" encoding="utf-8"?>
 <Project>
-  <!-- Enable Central Package Management -->
   <PropertyGroup>
     <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
   </PropertyGroup>
-  <!-- Central Package Versions -->
   <ItemGroup>
-    <!-- Test Framework Packages -->
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
     <PackageVersion Include="xunit.v3" Version="2.0.2" />
     <PackageVersion Include="xunit.v3.runner.inproc.console" Version="2.0.2" />
     <PackageVersion Include="xunit.runner.visualstudio" Version="3.0.0" />
     <PackageVersion Include="Moq" Version="4.20.72" />
     <PackageVersion Include="coverlet.collector" Version="6.0.2" />
-    <!-- System Packages - Aligned with FrooxEngine -->
-    <PackageVersion Include="System.Threading.Tasks.Extensions" Version="4.6.0" />
-    <PackageVersion Include="System.Memory" Version="4.6.0" />
-    <PackageVersion Include="System.Numerics.Vectors" Version="4.6.0" />
-    <PackageVersion Include="Microsoft.Bcl.AsyncInterfaces" Version="9.0.0" />
-    <PackageVersion Include="System.Runtime.CompilerServices.Unsafe" Version="6.1.0" />
-    <PackageVersion Include="System.Collections.Immutable" Version="9.0.0" />
     <PackageVersion Include="Resonite.GameLibs" Version="2025.9.2.430" />
   </ItemGroup>
 </Project>
